remove ShadowsocksConfig from model
