+++ src/electron/tunnel_store.ts
@@ -15,7 +15,7 @@
 import * as fs from 'fs';
 import * as path from 'path';
 
-import {ShadowsocksConfig} from '../www/model/shadowsocks';
+import {ShadowsocksConfig} from '../www/app/config';
 
 // Format to store a tunnel configuration.
 export interface SerializableTunnel {
